+++ packages/lib/src/propTransform/propTransform.ts
@@ -1,7 +1,7 @@
 import type { AnyDataModel } from "../dataModel/BaseDataModel"
 import type { AnyModel } from "../model/BaseModel"
 import { AnyModelProp, noDefaultValue } from "../modelShared/prop"
-import { addLateInitializationFunction, failure, runAfterNewSymbol } from "../utils"
+import { failure } from "../utils"
 
 /**
  * A prop transform.
@@ -25,85 +25,6 @@ export interface PropTransform<TProp, TData> {
    */
   dataToProp(dataValue: TData): TProp
 }
-
-/**
- * A prop transform decorator.
- */
-export type PropTransformDecorator<TProp> = <PK extends string>(
-  boundPropName: PK
-) => (target: { [k in PK]: TProp }, propertyKey: string) => void
-
-/**
- * Creates a prop transform, useful to transform property data into another kind of data.
- *
- * For example, to transform from a number timestamp into a date:
- * ```ts
- * const asDate = propTransform({
- *   propToData(prop: number) {
- *     return new Date(prop)
- *   },
- *   dataToProp(data: Date) {
- *     return +data
- *   }
- * })
- *
- * @model("myApp/MyModel")
- * class MyModel extends Model({
- *   timestamp: prop<number>()
- * }) {
- *   @asDate("timestamp")
- *   date!: Date
- * }
- *
- * const date: Date = myModel.date
- * // inside some model action
- * myModel.date = new Date()
- * ```
- *
- * @typename TProp Property value type.
- * @typename TData Data value type.
- * @param transform Transform to apply.
- * @returns A decorator that can be used to decorate model class fields.
- */
-export function propTransform<TProp, TData>(
-  transform: PropTransform<TProp, TData>
-): PropTransformDecorator<TProp> & typeof transform {
-  const parametrizedDecorator: PropTransformDecorator<TProp> = (boundPropName) => {
-    const decorator = (target: any, propertyKey: string) => {
-      addLateInitializationFunction(target, runAfterNewSymbol, (instance) => {
-        // make the field a getter setter
-        Object.defineProperty(instance, propertyKey, {
-          get(this: AnyModel | AnyDataModel): TData {
-            const memoTransform = memoTransformCache.getOrCreateMemoTransform(
-              this,
-              propertyKey,
-              transform
-            )
-
-            return memoTransform.propToData(this.$[boundPropName])
-          },
-          set(this: AnyModel | AnyDataModel, value: any) {
-            const memoTransform = memoTransformCache.getOrCreateMemoTransform(
-              this,
-              propertyKey,
-              transform
-            )
-
-            this.$[boundPropName] = memoTransform.dataToProp(value)
-            return true
-          },
-        })
-      })
-    }
-
-    return decorator
-  }
-  ;(parametrizedDecorator as any).propToData = transform.propToData.bind(transform)
-  ;(parametrizedDecorator as any).dataToProp = transform.dataToProp.bind(transform)
-
-  return parametrizedDecorator as any
-}
-
 class MemoTransformCache {
   private readonly cache = new WeakMap<object, Map<string, MemoPropTransform<any, any>>>()